Fully enumerate KSMv1 partial-emission metrics
The list of metrics we claimed to emit in the "only KSM v1" case was incomplete. This change fixes that list to fully enumerate emitted metrics in this case and links directly to the OSS code that is responsible for this emission for future debugging purposes.

We may wish to revisit which metrics exactly are emitted in this case to make sure it is the minimal set required.